feat : add image asset and add directory to separate `icons` and `images`
